# copilot-workspace-service/src/main/java/com/sporekart/workspace/dto/SwitchCopilotResponse.java
package com.sporekart.workspace.dto;

import java.util.Map;

public record SwitchCopilotResponse(
    String sessionId,
    String previousCopilotId,
    String currentCopilotId,
    String status,
    String message,
    Map<String, Object> context
) {}